Add middleware for user validation and role authorization, and refactor resource creation
