+++ index.js
@@ -3,10 +3,10 @@ const { dbConection } = require("./database/config");
 const cors = require("cors");
 
 const corsOptions = {
-    origin: "*",
-    credentials: true,
-    optionSuccessStatus: 200,
-    "Access-Control-Allow-Origin": "*",
+  origin: "*",
+  credentials: true,
+  optionSuccessStatus: 200,
+  "Access-Control-Allow-Origin": "*",
 };
 
 require("dotenv").config();
@@ -28,9 +28,10 @@ app.use(express.json());
 //todo: auth//create, login, renew del token
 app.use("/api/auth", require("./routes/auth"));
 app.use("/api/course", require("./routes/course"));
+app.use("/api/resources", require("./routes/resources"));
 //todo: CRUD: EVENTOS
 
 //listen requests
 app.listen(process.env.PORT, () => {
-    console.log(`Servidor corriendo en puerto ${process.env.PORT}`);
+  console.log(`Servidor corriendo en puerto ${process.env.PORT}`);
 });
